fix(tools): extract shared RevealInExplorer and ToolUploadHelper to eliminate code duplication
